server-python/requirements.txt
graphene==3.3
flask==3.0.0
flask-cors==4.0.0
flask-graphql==2.0.1
pyjwt==2.8.0
graphql-ws==0.4.4
